TX-7 clean_test_updateActivity

+++ src/main/java/com/example/adventurebackend/repository/ActivityRepository.java
@@ -3,4 +3,4 @@ package com.example.adventurebackend.repository;
 import com.example.adventurebackend.model.Activity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ActivityRepository extends JpaRepository<Activity, Long> {}
+public interface ActivityRepository extends JpaRepository<Activity, Integer> {}
